est_lp

+++ local_ic/1-Estimation_pente.R
@@ -4,17 +4,19 @@ library(rjdfilters)
 plan(multisession)
 if(!dir.exists("data_simul/byseriespente"))
   dir.create("data_simul/byseriespente")
-
+if(!dir.exists("data/byseriespente_nber"))
+  dir.create("data/byseriespente_nber")
 X_gen <- function(d = 1, p = 6, q = p){
   sapply(0:d, function(exp) seq(-p, q)^exp)
 }
+
 gen_MM <- function(p=6, q=p, d=2){
-  k = rjdfilters::get_kernel("Henderson", h = h)
+  k = rjdfilters::get_kernel("Henderson", h = p)
   k
   k = c(rev(k$coef[-1]), k$coef[seq(0,q)+1])
   k
   K = diag(k)
-  X = X_gen(d=d, p = h, q = q)
+  X = X_gen(d=d, p = p, q = q)
   e1 = e2 = e3 = matrix(0, ncol = 1, nrow = d+1)
   e1[1] = 1
   e2[2] = 1
@@ -38,7 +40,9 @@ d = 2
 for(s in list.files("data_simul/byseries",full.names = TRUE)){
   print(s)
   for(h in 3:6){
-    new_f = sprintf("data_simul/byseriespente/%s_h%i", basename(s),h)
+    new_f = sprintf("data_simul/byseriespente/%s_h%i.RDS",
+                    gsub(".RDS", "",basename(s)),h)
+    print(new_f)
     MM_h = MM[[sprintf("h=%i", h)]]
     if(!file.exists(new_f)){
       data <- readRDS(s)
@@ -48,17 +52,49 @@ for(s in list.files("data_simul/byseries",full.names = TRUE)){
           res3 = na.omit(jasym_filter(x, MM_h[[sprintf("d=%i",3)]],-h))
           resderiv2 = na.omit(jasym_filter(x, MM_h[["deriv2"]],-h))
           sigma2 =  mean((x - henderson(x,length = 2*h+1,musgrave = F))^2,
-                          na.rm = TRUE)
+                         na.rm = TRUE)
           sigma2 = sigma2/(1- 2*MM_h[["henderson"]]["t"] + sum(MM_h[["henderson"]]^2))
           names(sigma2) <- NULL
           c(`d=2` = res2[length(res2)],
             `d=3` = res3[length(res3)],
             `deriv2` = resderiv2[length(resderiv2)],
             `sigma2` = sigma2
-            )
+          )
         }
         )
-        })
+      })
+      info <- lapply(info_fs, value)
+      saveRDS(info, new_f)
+    }
+  }
+}
+
+
+for(s in list.files("data/byseries_nber",full.names = TRUE)){
+  print(s)
+  for(h in 3:6){
+    new_f = sprintf("data/byseriespente_nber/%s_h%i.RDS", gsub(".RDS", "",basename(s)),h)
+    print(new_f)
+    MM_h = MM[[sprintf("h=%i", h)]]
+    if(!file.exists(new_f)){
+      data <- readRDS(s)
+      info_fs <- lapply(data, function(x){
+        future({
+          res2 = na.omit(jasym_filter(x, MM_h[[sprintf("d=%i",2)]],-h))
+          res3 = na.omit(jasym_filter(x, MM_h[[sprintf("d=%i",3)]],-h))
+          resderiv2 = na.omit(jasym_filter(x, MM_h[["deriv2"]],-h))
+          sigma2 =  mean((x - henderson(x,length = 2*h+1,musgrave = F))^2,
+                         na.rm = TRUE)
+          sigma2 = sigma2/(1- 2*MM_h[["henderson"]]["t"] + sum(MM_h[["henderson"]]^2))
+          names(sigma2) <- NULL
+          c(`d=2` = res2[length(res2)],
+            `d=3` = res3[length(res3)],
+            `deriv2` = resderiv2[length(resderiv2)],
+            `sigma2` = sigma2
+          )
+        }
+        )
+      })
       info <- lapply(info_fs, value)
       saveRDS(info, new_f)
     }
